Purge non-UE legacy and dead code; fix shadowing and warnings
- Remove Leon-only legacy classes with no UE counterpart: bobbing, orbit and
  interactable components, VolumeHelpers, UMG interaction prompt and menu
  list widgets, AIChaseBehavior, BasicShape/BasicLight level helpers.
- APainCausingVolume now applies DamagePerSec * PainInterval to the pawns it
  encompasses every PainInterval (CausePainTo), and ticks (AVolume does not).
- AAIController exposes UE's EPathFollowingStatus / GetMoveStatus instead of
  Leon's logic state and wish direction API.
- Drop unused game mode, game state, player state, window, physics and mesh
  helpers; tangent golden reference moves to RenderCore test support.
- ShooterGame: remove unused blackboard keys, accessors and the menu stub.
- Rename locals and parameters that hid inherited members (MSVC C4458,
  GCC -Wshadow): fixes the Win64 CI build.
- Linux warnings aligned with MSVC (-Wall -Wextra -Werror=shadow); fix the
  sign-compare, the Shipping-only unused parameter and the FString append
  range the optimizer could not prove.
- Fix comments that described removed or non-existent behavior.
- Test: System.Engine.Damage.PainCausingVolume.

+++ Engine/Source/Runtime/Engine/Private/SceneView.cpp
@@ -12,7 +12,7 @@ namespace
 bool* FEngineShowFlags::FindFlag(const FString& Name)
 {
 	bool* const Flags[] = {&Bounds, &Collision, &Navigation, &AxesGizmo};
-	for (int32 Index = 0; Index < UE_ARRAY_COUNT(ShowFlagNames); ++Index)
+	for (int32 Index = 0; Index < int32(UE_ARRAY_COUNT(ShowFlagNames)); ++Index)
 	{
 		if (Name.Equals(ShowFlagNames[Index], ESearchCase::IgnoreCase))
 		{
